Respawn agents when they collide with enemies
If agents from opposing teams try to occupy the same space, they will be sent
back to their spawn point. When multiple agents from multiple teams collide, the
team with the most agents gets to keep the difference in its count compared to
the team with the second most. All others are respawned.