Add completion criteria
The `Victory` classes are responsible for knowing when the game is over and who
has won. They must implement `#score`, and probably want to implement `#done?`
as well.

We may want to update players' scores each tick to pass as information to their
`Agent`s or just show on a leaderboard during games.